Tidies the `Disk` file hierarchy.
