+++ Storage/Disk/Controller/DiskController.cpp
@@ -8,7 +8,7 @@
 
 #include "DiskController.hpp"
 
-#include "../../NumberTheory/Factors.hpp"
+#include "../../../NumberTheory/Factors.hpp"
 
 using namespace Storage::Disk;
