+++ Storage/Disk/DiskImage/Formats/CPCDSK.hpp
@@ -10,7 +10,7 @@
 #define CPCDSK_hpp
 
 #include "../DiskImage.hpp"
-#include "../../FileHolder.hpp"
+#include "../../../FileHolder.hpp"
 
 #include <vector>
