+++ Storage/Disk/DiskImage/Formats/HFE.cpp
@@ -8,7 +8,7 @@
 
 #include "HFE.hpp"
 
-#include "../PCMTrack.hpp"
+#include "../../Track/PCMTrack.hpp"
 
 using namespace Storage::Disk;
