+++ Storage/Disk/DiskImage/Formats/HFE.hpp
@@ -10,7 +10,7 @@
 #define HFE_hpp
 
 #include "../DiskImage.hpp"
-#include "../../FileHolder.hpp"
+#include "../../../FileHolder.hpp"
 
 namespace Storage {
 namespace Disk {
