+++ Storage/Disk/Track/PCMSegment.hpp
@@ -12,8 +12,8 @@
 #include <cstdint>
 #include <vector>
 
-#include "../Storage.hpp"
-#include "Disk.hpp"
+#include "../../Storage.hpp"
+#include "Track.hpp"
 
 namespace Storage {
 namespace Disk {
